fix: Show correct map data on show live for trip
