+++ src/travel-details-screens/components/TripSection.tsx
@@ -51,6 +51,8 @@ import {
 import {usePreferences} from '@atb/preferences';
 import {Button} from '@atb/components/button';
 import {Map} from '@atb/assets/svg/mono-icons/map';
+import {ServiceJourneyMapInfoData_v3} from '@atb/api/types/serviceJourney';
+import {useMapData} from '@atb/travel-details-screens/use-map-data';
 
 type TripSectionProps = {
   isLast?: boolean;
@@ -60,7 +62,7 @@ type TripSectionProps = {
   interchangeDetails?: InterchangeDetails;
   leg: Leg;
   testID?: string;
-  onPressShowLive?(): void;
+  onPressShowLive?(mapData: ServiceJourneyMapInfoData_v3): void;
   onPressDeparture: TripProps['onPressDeparture'];
   onPressQuay: TripProps['onPressQuay'];
 };
@@ -108,6 +110,12 @@ export const TripSection: React.FC<TripSectionProps> = ({
   const lastPassedStop = leg.serviceJourneyEstimatedCalls
     ?.filter((a) => !a.predictionInaccurate && a.actualDepartureTime)
     .pop();
+
+  const mapData = useMapData(
+    leg.serviceJourney?.id,
+    leg.fromPlace.quay?.id,
+    leg.toPlace.quay?.id,
+  );
 
   const bookingDetails: ContactDetails | undefined = leg?.bookingArrangements
     ?.bookingContact?.phone &&
@@ -225,14 +233,14 @@ export const TripSection: React.FC<TripSectionProps> = ({
             />
           </TripRow>
         )}
-        {onPressShowLive ? (
+        {onPressShowLive && mapData ? (
           <TripRow>
             <Button
               type="pill"
               leftIcon={{svg: Map}}
               text={t(TripDetailsTexts.trip.leg.live)}
               interactiveColor="interactive_3"
-              onPress={onPressShowLive}
+              onPress={() => onPressShowLive(mapData)}
             />
           </TripRow>
         ) : null}
